基礎アルゴリズム作成
