+++ src/get_production_data/get_param/get_quantity.py
@@ -32,18 +32,3 @@ def Get_quantity(d_soup):
       quantity = 0
       return quantity
 
-def Get_quantity_indexPage(data):
-  try:
-    stock_text = data.select(".a-color-price", recursive=False)[0].get_text()
-    if "残り" in stock_text:
-      if "点" in stock_text[3]:
-        stock = int(stock_text[2])
-      else:
-        stock = int(stock_text[2:4])
-    else:
-      stock = 100
-    return stock
-
-  except IndexError:
-    stock = 100
-    return stock
